requirements.txt

streamlit
youtube-transcript-api
transformers
torch
fpdf
python-docx
openai
python-dotenv
google-generativeai